Remove old obsolete fixture
