+++ tests/SqlStreamStore.MsSql.V3.Tests/MsSqlStreamStoreV3AcceptanceTests.cs
@@ -13,14 +13,14 @@
         { }
 
         protected override async Task<IStreamStoreFixture> CreateFixture()
-            => await MsSqlStreamStoreV3Fixture2.Create("foo");
+            => await MsSqlStreamStoreV3Fixture.Create("foo");
 
         [Fact]
         public async Task Can_use_multiple_schemas()
         {
             var dboStore = fixture.Store;
 
-            using (var barFixture = await MsSqlStreamStoreV3Fixture2.Create("bar"))
+            using (var barFixture = await MsSqlStreamStoreV3Fixture.Create("bar"))
             {
                 var barStore = barFixture.Store;
 
@@ -42,7 +42,7 @@
         [Theory, InlineData("dbo"), InlineData("myschema")]
         public async Task Can_call_initialize_repeatably(string schema)
         {
-            using (var fixture = await MsSqlStreamStoreV3Fixture2.Create(schema, createSchema:false))
+            using (var fixture = await MsSqlStreamStoreV3Fixture.Create(schema, createSchema:false))
             {
                 await fixture.Store.CreateSchemaIfNotExists();
                 await fixture.Store.CreateSchemaIfNotExists();
@@ -52,7 +52,7 @@
         [Fact]
         public async Task Can_drop_all()
         {
-            using (var fixture = await MsSqlStreamStoreV3Fixture2.Create())
+            using (var fixture = await MsSqlStreamStoreV3Fixture.Create())
             {
                 await fixture.Store.DropAll();
             }
@@ -61,7 +61,7 @@
         [Fact]
         public async Task Can_check_schema()
         {
-            using (var fixture = await MsSqlStreamStoreV3Fixture2.Create())
+            using (var fixture = await MsSqlStreamStoreV3Fixture.Create())
             {
                 var result = await fixture.Store.CheckSchema();
 
@@ -74,7 +74,7 @@
         [Fact]
         public async Task When_schema_is_not_created_then_should_be_indicated()
         {
-            using (var fixture = await MsSqlStreamStoreV3Fixture2.Create(createSchema: false))
+            using (var fixture = await MsSqlStreamStoreV3Fixture.Create(createSchema: false))
             {
                 var result = await fixture.Store.CheckSchema();
